refactor(app): follow the renderer split's single-consumer items to their callers
Step 0 of the review found three cleanups and no defect: seven items in
footprint_render.rs that only bar.rs reads (the imbalance, chip, badge and
backdrop constants, canvas_backdrop and row_band) move to bar.rs; the three
orderflow_render root helpers with one sibling caller move to it
(readable_band and finite_unit_f64 to layout.rs, draw_text_with_shadow to
heatmap.rs); and the test sidecar's header, which still said the move widened
nothing, now names what is pub(super) and why. Every body is unchanged; the
statement-line multiset proof is the same as before the batch.
