+++ crates/app/src/orderflow_render/tests/mod.rs
@@ -3,9 +3,14 @@
 // for.
 //
 // They stay a child module of `crate::orderflow_render` rather than moving to an
-// integration test: a child sees its ancestor's private items, so the move
-// widens no visibility in production code, and the `use super::*` below is
-// the line the module already had inline.
+// integration test: a child sees its ancestor's private items and, through
+// `use super::*`, the root's own. The renderer has since been split into
+// sibling modules, and an item a test names in one of them is `pub(super)`
+// so this sidecar can still reach it - `bubbles::crown_geometry`,
+// `heatmap::gap_marks`, `legend::flow_layout` and their like have no
+// production caller outside their own file and carry that visibility for
+// these tests alone. The five glob imports below are how the sidecar sees
+// them; nothing here is `pub(crate)` or wider.
 
 use super::bubbles::*;
 use super::heatmap::*;
